ref(server): Move all managed types into a dedicated managed module
