+++ relay-server/src/lib.rs
@@ -258,6 +258,7 @@ mod endpoints;
 mod envelope;
 mod extractors;
 mod http;
+mod managed;
 mod metrics;
 mod metrics_extraction;
 mod middlewares;
