+++ relay-server/src/managed/mod.rs
@@ -1,0 +1,13 @@
+mod counted;
+mod envelope;
+#[expect(
+    clippy::module_inception,
+    reason = "private module, containing the Managed type, module exists to prevent clashes with the envelope module"
+)]
+mod managed;
+mod utils;
+
+pub use self::counted::*;
+pub use self::envelope::*;
+pub use self::managed::*;
+pub use self::utils::*;
