+++ relay-server/src/services/processor/session.rs
@@ -15,10 +15,10 @@ use relay_metrics::Bucket;
 use relay_statsd::metric;
 
 use crate::envelope::{ContentType, Item, ItemType};
+use crate::managed::{ItemAction, TypedEnvelope};
 use crate::services::processor::{MINIMUM_CLOCK_DRIFT, ProcessingExtractedMetrics, SessionGroup};
 use crate::services::projects::project::ProjectInfo;
 use crate::statsd::RelayTimers;
-use crate::utils::{ItemAction, TypedEnvelope};
 
 #[derive(Debug, Clone, Copy)]
 struct SessionProcessingConfig<'a> {
